Fix TS codegen circular import issue
When a message is inherited, the class definition contains a static factory function that takes Uint8Array and creates the correct concrete message. Doing so requires importing all class definitions of child messages in order to have access to their factory function. Since the class definitions of child messages also require the import of the parent message to extend the parent message class, this creates a circular import. To circumvent this, the factory function of the parent message that creates concrete message types is moved to a separate file.